+++ test/heat_conduction_1D.jl
@@ -41,7 +41,7 @@ Tests for 1D heat conduction solvers:
     n_depth = 101  # Number of depth steps
     Δz = z_max / (n_depth - 1)  # Depth step width [m]
 
-    thermo_params = AsteroidThermoPhysicalModels.ThermoParams(k, ρ, Cₚ, R_vis, R_ir, ε, z_max, Δz, n_depth)
+    thermo_params = ThermoParams(k, ρ, Cₚ, R_vis, R_ir, ε, z_max, Δz, n_depth)
 
     ## --- TPMs with different solvers ---
     SELF_SHADOWING = false